--- components/Main/youtube.module.css
@@ -1,12 +1,0 @@
-.event {
-    margin-top: 40px;
-    margin-bottom: 100px;
-    text-decoration: none;
-
-}
-
-.event a {
-
-    text-decoration: none;
-
-}
