--- components/Navbar/Navbar.module.css
@@ -1,139 +1,0 @@
-@font-face {
-  font-family: 'Hahmlet-Regular';
-  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2110@1.0/Hahmlet-Regular.woff2')
-    format('woff2');
-  font-weight: normal;
-  font-style: normal;
-}
-
-.navbar {
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 80px;
-  line-height: 80px;
-  background: linear-gradient(
-    to bottom,
-    rgba(255, 255, 255),
-    rgba(235, 235, 235)
-  );
-  z-index: 1000;
-  display: flex;
-  font-family: 'Hahmlet-Regular';
-  font-weight: 1000;
-  font-size: 18px;
-}
-
-.logo {
-  width: 300px;
-  height: 80px;
-  text-align: center;
-  margin: 0; /* 마진 제거 */
-}
-
-.liInline {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  /* margin-left: 200px; */
-}
-
-.liInline > li {
-  display: inline-block;
-  width: 200px;
-  height: 80px;
-  text-align: center;
-}
-
-.liInline > input {
-  width: 200px;
-  height: 28px;
-  vertical-align: middle;
-  margin-left: 10px;
-  margin-bottom: 6px;
-}
-.liInline > button {
-  width: 28px;
-  height: 28px;
-  vertical-align: middle;
-  margin-left: 10px;
-  margin-bottom: 6px;
-  cursor: pointer;
-  border: none;
-  background: transparent;
-}
-
-.b {
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.dropdownWrapper {
-  position: relative;
-  display: inline-block;
-  cursor: pointer;
-}
-
-.dropdownMenu {
-  display: none;
-  position: absolute;
-  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-  background-color: rgba(235, 245, 245, 0.9);
-  width: 100%;
-}
-
-.dropdownMenu ul {
-  list-style: none;
-  padding: 0;
-}
-
-.dropdownWrapper:hover .dropdownMenu {
-  display: block;
-}
-
-.dropdownWrapper:hover {
-  color: rgba(100, 100, 255, 0.9);
-}
-
-.p {
-  color: black;
-  text-decoration: none;
-}
-
-.p:hover {
-  color: rgba(100, 100, 255, 0.9);
-}
-
-.link {
-  position: relative;
-  display: inline-block;
-  text-decoration: none;
-  overflow: hidden;
-  color: black;
-}
-
-.link::before {
-  content: '';
-  position: absolute;
-  width: 0;
-  height: 2px;
-  bottom: 15px;
-  left: 50%;
-  background-color: rgba(100, 100, 255, 0.9);
-  transition: all 0.3s ease;
-}
-
-.link:hover::before {
-  width: 100%;
-  left: 0;
-}
-
-.user {
-  width: 300px;
-  height: 80px;
-  text-align: center;
-  line-height: 30px; /* 텍스트 수직 가운데 정렬 */
-  color: black;
-}
