--- components/Sidebar/Sidebar.module.css
@@ -1,49 +1,0 @@
-.menuButton {
-  position: fixed;
-  top: 50%;
-  left: 0;
-  transition: transform 0.48s ease-in-out;
-}
-
-.sidebar {
-  width: 250px;
-  position: fixed;
-  top: 80px;
-  left: -250px;
-  /* 초기 상태에서 사이드바가 화면 밖에 위치하도록 설정 */
-  height: 100%;
-  background-color: #223355;
-  transition: transform 0.5s ease-in-out;
-  /* 부드러운 애니메이션 효과를 위해 transition 속성 추가 */
-  display: flex;
-  flex-direction: column;
-  /* 메뉴 항목을 세로로 나열 */
-  padding-top: 60px;
-  /* 상단에 공간 추가 */
-}
-
-.sidebarLink {
-  padding: 25px;
-  color: white;
-  font-size: 18px;
-  text-align: left;
-  margin-bottom: 50px;
-  background-color: #223355;
-  transition: background-color 0.3s;
-  border: none;
-}
-
-.sidebarLink:hover {
-  background-color: RGB(102, 204, 255);
-  /* 마우스 호버 시 배경색 변경 */
-}
-
-.sidebarOpen {
-  transform: translateX(250px);
-  transition: transform 0.5s ease-in-out;
-  /* 사이드바가 열릴 때 오른쪽에서 0의 위치로 이동 */
-}
-
-.icon {
-  margin-right: 20px;
-}
