--- components/Sidebar/Traffic.module.css
@@ -1,21 +1,0 @@
-.circle {
-  fill: red; /* 원의 색상 */
-  stroke: black; /* 원의 테두리 색상 */
-  stroke-width: 1; /* 원의 테두리 두께 */
-}
-
-.line {
-  stroke: blue; /* 선의 색상 */
-  stroke-width: 2; /* 선의 두께 */
-}
-
-.gridLine {
-  stroke: rgba(0, 0, 0, 0.2); /* 격자 선의 색상 */
-  stroke-width: 1; /* 격자 선의 두께 */
-}
-.graph {
-  text-align: center;
-}
-.label {
-  margin-left: 300px;
-}
