+++ src/app/components/Main/soga.module.css
