+++ src/app/components/Main/soga2.module.css
