+++ src/app/components/Main/soga3.module.css
