+++ src/app/components/Main/soga3.tsx
@@ -1,13 +1,13 @@
-import React from 'react';
-import styles from './soga2.module.css';
-import Image from 'next/image';
+import React from "react";
+import styles from "./soga3.module.css";
+import Image from "next/image";
 
-const Soga2 = () => {
+const Soga3 = () => {
   return (
     <div className={styles.container}>
       <div className={styles.imageContainer}>
         <Image
-          src="/pic.png"
+          src="/assets/pic2.png"
           alt="hifive img"
           width={500}
           height={330}
@@ -15,10 +15,10 @@ const Soga2 = () => {
         />
       </div>
       <div className={styles.description}>
-        공격이 감지되면 자동으로 트래픽을 필터링하며 차단
+        공격 시도와 시스템 반응에 대한 상세한 로그를 제공
       </div>
     </div>
   );
 };
 
-export default Soga2;
+export default Soga3;
